split apart table file
