+++ lib/rets/metadata.rb
@@ -4,3 +4,5 @@ require 'rets/metadata/resource'
 require 'rets/metadata/rets_class'
 require 'rets/metadata/root'
 require 'rets/metadata/table'
+require 'rets/metadata/lookup_table'
+require 'rets/metadata/table_factory'
